feat: pos block number calc
